requirements.txt
asgiref>=3.3.1
Django>=3.1.7
pytz>=2021.1
sqlparse>=0.4.1
mysqlclient>=2.0
djangorestframework>=3.14.0
django_mysql>=4.0
redis>=4.4.5
pytest
pytest-django
drf_spectacular